Clearned up store
